+++ multibus/__init__.py
@@ -16,7 +16,7 @@
 #
 """Multi-Bus Gateway package."""
 
-__version__ = "3.1.3"
+__version__ = "3.1.4"
 __author__ = "sm26449"
 
 from .config import Config